Move some variant analysis manager test to activated extension suite
