+++ extensions/ql-vscode/test/vscode-tests/activated-extension/jest.setup.ts
@@ -1,4 +1,5 @@
 import {
+  afterAllAction,
   beforeAllAction,
   beforeEachAction,
 } from "../jest.activated-extension.setup";
@@ -9,4 +10,8 @@ beforeAll(async () => {
 
 beforeEach(async () => {
   await beforeEachAction();
+});
+
+afterAll(async () => {
+  await afterAllAction();
 });
